requirements.txt
requests == 2.31.0
numpy==1.21.0
scikit-learn==0.24.2
cartesi-wallet==0.2.0
cartesi-rollup==0.2.0
